Refactor parquet tests
