+++ velox/dwio/parquet/tests/CMakeLists.txt
@@ -15,6 +15,8 @@
 set(TEST_LINK_LIBS
     velox_dwio_common_test_utils
     velox_vector_test_lib
+    velox_exec_test_lib
+    velox_temp_path
     gtest
     gtest_main
     gmock
