+++ velox/dwio/parquet/tests/reader/ParquetPageReaderTest.cpp
@@ -15,16 +15,14 @@
  */
 
 #include "velox/dwio/parquet/reader/PageReader.h"
-#include "velox/dwio/parquet/reader/ParquetReader.h"
-#include "velox/dwio/parquet/tests/ParquetReaderTestBase.h"
+#include "velox/dwio/parquet/tests/ParquetTestBase.h"
 
 using namespace facebook::velox;
 using namespace facebook::velox::common;
 using namespace facebook::velox::dwio::common;
-using namespace facebook::velox::dwio::parquet;
 using namespace facebook::velox::parquet;
 
-class ParquetPageReaderTest : public ParquetReaderTestBase {};
+class ParquetPageReaderTest : public ParquetTestBase {};
 
 namespace {
 auto defaultPool = memory::addDefaultLeafMemoryPool();
